datapeace/requirements.txt
Django==2.2    
djangorestframework==3.9.2  
djangorestframework-stubs==0.4.2  
pip==10.0.1 
pytz==2019.1 
setuptools==39.1.0 
sqlparse==0.3.0  
